+++ CHANGELOG.md
@@ -4,6 +4,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Turtle (scute) material. Can be used to craft helmets only.
+- Turtle trait. Mimics the vanilla turtle shell helmet's effect.
+### Changed
+- Stats can now be specified per gear type. For instance, this allows the armor stat to be manually set for each piece, rather than using a predefined ratio. Old files should still work as before.
+- Netherite-coated armor now uses the vanilla model texture
 ### Fixed
 - Potion effect traits crashing if arrays are missing elements